moved onboarding to helper file
